Issue2_openDB/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
aqtp==0.7.4
astunparse==1.6.3
Authlib==1.3.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
cloudpickle==3.0.0
clu @ git+https://github.com/google/CommonLoopUtils@92334e9d7e8bd004f3ab8ccd596ea980aa86c364
contextlib2==21.6.0
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
decorator==5.1.1
distrax==0.1.5
distro==1.9.0
dm-tree==0.1.8
docstring_parser==0.16
editdistance==0.8.1
einops==0.8.0
environs==9.5.0
etils==1.7.0
exceptiongroup==1.2.1
filelock==3.14.0
flatbuffers==24.3.25
flax==0.8.4
flaxformer @ git+https://github.com/google/flaxformer@399ea3a85e9807ada653fd0de1a9de627eb0acde
fonttools==4.53.0
fsspec==2024.6.0
ftfy==6.2.0
gast==0.5.4
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
grpcio-health-checking==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
immutabledict==4.2.0
importlib_resources==6.4.0
jax==0.4.28
jaxlib==0.4.28
Jinja2==3.1.4
joblib==1.4.2
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
mdurl==0.1.2
milvus-lite==2.4.7
ml-collections==0.1.1
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
open-clip-torch==2.24.0
openai==1.34.0
opt-einsum==3.3.0
optax==0.2.2
optree==0.11.0
orbax-checkpoint==0.5.15
overrides==7.7.0
packaging==24.0
pandas==2.2.2
panopticapi @ git+https://github.com/akolesnikoff/panopticapi.git@a698a12deb21e4cf0f99ef0581b2c30c466bf355
pgvector==0.2.5
pillow==10.3.0
pinecone-client==4.1.1
pinecone-plugin-interface==0.0.7
promise==2.3
protobuf==5.27.1
psutil==5.9.8
psycopg2-binary==2.9.9
pyarrow==16.1.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pymilvus==2.4.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
simple_parsing==0.1.5
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-cpu==2.16.1
tensorflow-gan==2.1.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-metadata==1.15.0
tensorflow-probability==0.24.0
tensorflow-text==2.16.1
tensorstore==0.1.61
termcolor==2.4.0
tf_keras==2.16.0
tfds-nightly==4.9.6.dev202406110044
threadpoolctl==3.5.0
timm==1.0.3
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
transformers==4.41.2
typing_extensions==4.12.1
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
validators==0.28.3
wcwidth==0.2.13
weaviate-client==4.6.4
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
